Move logger to lib
